chore: update cert validator lambda for new cert
