+++ cf-custom-resources/test/dns-cert-validator-test.js
@@ -12,8 +12,23 @@ describe("DNS Validated Certificate Handler", () => {
 
   let origLog = console.log;
   const testRequestId = "f4ef1b10-c39a-44e3-99c0-fbf7e53c3943";
-  const testDomainName = "test.example.com";
-  const testHostedZoneId = "/hostedzone/Z3P5QSUBK4POTI";
+  const testAppName = "myapp";
+  const testEnvName = "test";
+  const testDomainName = "example.com";
+  const testHostedZoneId = "Z3P5QSUBK4POTI";
+  const testAppHostedZoneId = "K4POTIZ3P5QSUB";
+  const testRootDNSRole = "mockRole";
+  const testSubjectAlternativeNames = [
+    "example.com",
+    "*.example.com",
+    "myapp.example.com",
+    "*.myapp.example.com",
+    "*.test.myapp.example.com",
+  ];
+  const testCopilotTags = [
+    { Key: "copilot-application", Value: testAppName },
+    { Key: "copilot-environment", Value: testEnvName },
+  ];
   const testCertificateArn =
     "arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012";
   const testRRName = "_3639ac514e785e898d2646601fa951d5.example.com";
@@ -21,6 +36,80 @@ describe("DNS Validated Certificate Handler", () => {
   const spySleep = sinon.spy(function (ms) {
     return Promise.resolve();
   });
+  const testDeleteRecordChangebatch = {
+    Changes: [
+      {
+        Action: "DELETE",
+        ResourceRecordSet: {
+          Name: testRRName,
+          Type: "CNAME",
+          TTL: 60,
+          ResourceRecords: [
+            {
+              Value: testRRValue,
+            },
+          ],
+        },
+      },
+    ],
+  };
+  const testUpsertRecordChangebatch = {
+    Changes: [
+      {
+        Action: "UPSERT",
+        ResourceRecordSet: {
+          Name: testRRName,
+          Type: "CNAME",
+          TTL: 60,
+          ResourceRecords: [
+            {
+              Value: testRRValue,
+            },
+          ],
+        },
+      },
+    ],
+  };
+  const legacyCertValidatorOptions = [
+    {
+      DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+      ValidationStatus: "SUCCESS",
+      ResourceRecord: {
+        Name: testRRName,
+        Type: "CNAME",
+        Value: testRRValue,
+      },
+    },
+  ];
+  const newCertValidateOptions = [
+    {
+      DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+      ValidationStatus: "SUCCESS",
+      ResourceRecord: {
+        Name: testRRName,
+        Type: "CNAME",
+        Value: testRRValue,
+      },
+    },
+    {
+      DomainName: `${testAppName}.${testDomainName}`,
+      ValidationStatus: "SUCCESS",
+      ResourceRecord: {
+        Name: testRRName,
+        Type: "CNAME",
+        Value: testRRValue,
+      },
+    },
+    {
+      DomainName: `${testDomainName}`,
+      ValidationStatus: "SUCCESS",
+      ResourceRecord: {
+        Name: testRRName,
+        Type: "CNAME",
+        Value: testRRValue,
+      },
+    },
+  ];
 
   beforeEach(() => {
     handler.withDefaultResponseURL(ResponseURL);
@@ -80,7 +169,7 @@ describe("DNS Validated Certificate Handler", () => {
       });
   });
 
-  test("Create operation requests a certificate", () => {
+  test("Create operation requests a new certificate", () => {
     const requestCertificateFake = sinon.fake.resolves({
       CertificateArn: testCertificateArn,
     });
@@ -94,16 +183,110 @@ describe("DNS Validated Certificate Handler", () => {
     describeCertificateFake.resolves({
       Certificate: {
         CertificateArn: testCertificateArn,
-        DomainValidationOptions: [
-          {
-            ValidationStatus: "SUCCESS",
-            ResourceRecord: {
-              Name: testRRName,
-              Type: "CNAME",
-              Value: testRRValue,
-            },
-          },
-        ],
+        DomainValidationOptions: newCertValidateOptions,
+      },
+    });
+
+    const changeResourceRecordSetsFake = sinon.fake.resolves({
+      ChangeInfo: {
+        Id: "bogus",
+      },
+    });
+
+    const listHostedZonesByNameFake = sinon.fake.resolves({
+      HostedZones: [
+        {
+          Id: `/hostedzone/${testHostedZoneId}`,
+        },
+      ],
+    });
+
+    AWS.mock("ACM", "requestCertificate", requestCertificateFake);
+    AWS.mock("ACM", "describeCertificate", describeCertificateFake);
+    AWS.mock(
+      "Route53",
+      "changeResourceRecordSets",
+      changeResourceRecordSetsFake
+    );
+    AWS.mock("Route53", "listHostedZonesByName", listHostedZonesByNameFake);
+
+    const request = nock(ResponseURL)
+      .put("/", (body) => {
+        return body.Status === "SUCCESS";
+      })
+      .reply(200);
+
+    return LambdaTester(handler.certificateRequestHandler)
+      .event({
+        RequestType: "Create",
+        RequestId: testRequestId,
+        ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
+          IsAliasEnabled: "true",
+          Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+          SubjectAlternativeNames: testSubjectAlternativeNames,
+        },
+      })
+      .expectResolve(() => {
+        sinon.assert.calledWith(
+          describeCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        sinon.assert.calledWith(
+          requestCertificateFake,
+          sinon.match({
+            DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+            SubjectAlternativeNames: testSubjectAlternativeNames,
+            ValidationMethod: "DNS",
+            Tags: testCopilotTags,
+          })
+        );
+        sinon.assert.calledWith(
+          changeResourceRecordSetsFake,
+          sinon.match({
+            ChangeBatch: testUpsertRecordChangebatch,
+            HostedZoneId: testHostedZoneId,
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testAppName}.${testDomainName}`,
+            MaxItems: "1",
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testDomainName}`,
+            MaxItems: "1",
+          })
+        );
+        expect(request.isDone()).toBe(true);
+      });
+  });
+
+  test("Create operation requests a legacy certificate", () => {
+    const requestCertificateFake = sinon.fake.resolves({
+      CertificateArn: testCertificateArn,
+    });
+
+    const describeCertificateFake = sinon.stub();
+    describeCertificateFake.onFirstCall().resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+      },
+    });
+    describeCertificateFake.resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+        DomainValidationOptions: legacyCertValidatorOptions,
       },
     });
 
@@ -132,39 +315,38 @@ describe("DNS Validated Certificate Handler", () => {
         RequestType: "Create",
         RequestId: testRequestId,
         ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
           DomainName: testDomainName,
-          HostedZoneId: testHostedZoneId,
+          EnvHostedZoneId: testHostedZoneId,
+          IsAliasEnabled: "false",
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+          SubjectAlternativeNames: testSubjectAlternativeNames,
         },
       })
       .expectResolve(() => {
         sinon.assert.calledWith(
+          describeCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        sinon.assert.calledWith(
           requestCertificateFake,
           sinon.match({
-            DomainName: testDomainName,
+            DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+            SubjectAlternativeNames: [
+              `*.${testEnvName}.${testAppName}.${testDomainName}`,
+            ],
             ValidationMethod: "DNS",
+            Tags: testCopilotTags,
           })
         );
         sinon.assert.calledWith(
           changeResourceRecordSetsFake,
           sinon.match({
-            ChangeBatch: {
-              Changes: [
-                {
-                  Action: "UPSERT",
-                  ResourceRecordSet: {
-                    Name: testRRName,
-                    Type: "CNAME",
-                    TTL: 60,
-                    ResourceRecords: [
-                      {
-                        Value: testRRValue,
-                      },
-                    ],
-                  },
-                },
-              ],
-            },
+            ChangeBatch: testUpsertRecordChangebatch,
             HostedZoneId: testHostedZoneId,
           })
         );
@@ -182,6 +364,7 @@ describe("DNS Validated Certificate Handler", () => {
       Certificate: {
         CertificateArn: testCertificateArn,
       },
+      DomainValidationOptions: newCertValidateOptions,
     });
 
     AWS.mock("ACM", "requestCertificate", requestCertificateFake);
@@ -203,16 +386,22 @@ describe("DNS Validated Certificate Handler", () => {
         RequestType: "Create",
         RequestId: testRequestId,
         ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
           DomainName: testDomainName,
-          HostedZoneId: testHostedZoneId,
+          EnvHostedZoneId: testHostedZoneId,
+          IsAliasEnabled: "true",
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+          SubjectAlternativeNames: testSubjectAlternativeNames,
         },
       })
       .expectResolve(() => {
         sinon.assert.calledWith(
           requestCertificateFake,
           sinon.match({
-            DomainName: testDomainName,
+            DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+            SubjectAlternativeNames: testSubjectAlternativeNames,
             ValidationMethod: "DNS",
           })
         );
@@ -256,16 +445,21 @@ describe("DNS Validated Certificate Handler", () => {
         RequestType: "Create",
         RequestId: testRequestId,
         ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
           DomainName: testDomainName,
-          HostedZoneId: testHostedZoneId,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+          SubjectAlternativeNames: testSubjectAlternativeNames,
         },
       })
       .expectResolve(() => {
         sinon.assert.calledWith(
           requestCertificateFake,
           sinon.match({
-            DomainName: testDomainName,
+            DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+            SubjectAlternativeNames: testSubjectAlternativeNames,
             ValidationMethod: "DNS",
           })
         );
@@ -276,6 +470,97 @@ describe("DNS Validated Certificate Handler", () => {
           .map((call) => call.args[0])
           .reduce((p, n) => p + n, 0);
         expect(totalSleep).toBeLessThan(360 * 1000);
+      });
+  });
+
+  test("Create operation fails because no hosted zone found", () => {
+    const requestCertificateFake = sinon.fake.resolves({
+      CertificateArn: testCertificateArn,
+    });
+
+    const describeCertificateFake = sinon.stub();
+    describeCertificateFake.onFirstCall().resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+      },
+    });
+    describeCertificateFake.resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+        DomainValidationOptions: [
+          {
+            DomainName: `${testAppName}.${testDomainName}`,
+            ValidationStatus: "SUCCESS",
+            ResourceRecord: {
+              Name: testRRName,
+              Type: "CNAME",
+              Value: testRRValue,
+            },
+          },
+        ],
+      },
+    });
+
+    const changeResourceRecordSetsFake = sinon.fake.resolves({
+      ChangeInfo: {
+        Id: "bogus",
+      },
+    });
+
+    const listHostedZonesByNameFake = sinon.fake.resolves({
+      HostedZones: [],
+    });
+
+    AWS.mock("ACM", "requestCertificate", requestCertificateFake);
+    AWS.mock("ACM", "describeCertificate", describeCertificateFake);
+    AWS.mock(
+      "Route53",
+      "changeResourceRecordSets",
+      changeResourceRecordSetsFake
+    );
+    AWS.mock("Route53", "listHostedZonesByName", listHostedZonesByNameFake);
+
+    const request = nock(ResponseURL)
+      .put("/", (body) => {
+        return (
+          body.Status === "FAILED" &&
+          body.Reason.startsWith("Couldn't find any Hosted Zone")
+        );
+      })
+      .reply(200);
+
+    return LambdaTester(handler.certificateRequestHandler)
+      .event({
+        RequestType: "Create",
+        RequestId: testRequestId,
+        ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
+          IsAliasEnabled: "true",
+          Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+          SubjectAlternativeNames: testSubjectAlternativeNames,
+        },
+      })
+      .expectResolve(() => {
+        sinon.assert.calledWith(
+          requestCertificateFake,
+          sinon.match({
+            DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+            SubjectAlternativeNames: testSubjectAlternativeNames,
+            ValidationMethod: "DNS",
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testAppName}.${testDomainName}`,
+            MaxItems: "1",
+          })
+        );
+        expect(request.isDone()).toBe(true);
       });
   });
 
@@ -290,16 +575,7 @@ describe("DNS Validated Certificate Handler", () => {
     const describeCertificateFake = sinon.fake.resolves({
       Certificate: {
         CertificateArn: testCertificateArn,
-        DomainValidationOptions: [
-          {
-            ValidationStatus: "SUCCESS",
-            ResourceRecord: {
-              Name: testRRName,
-              Type: "CNAME",
-              Value: testRRValue,
-            },
-          },
-        ],
+        DomainValidationOptions: legacyCertValidatorOptions,
       },
     });
     AWS.mock("ACM", "describeCertificate", describeCertificateFake);
@@ -326,9 +602,14 @@ describe("DNS Validated Certificate Handler", () => {
         RequestType: "Create",
         RequestId: testRequestId,
         ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
           DomainName: testDomainName,
-          HostedZoneId: testHostedZoneId,
+          EnvHostedZoneId: testHostedZoneId,
+          IsAliasEnabled: "true",
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+          SubjectAlternativeNames: testSubjectAlternativeNames,
         },
       })
       .expectResolve(() => {
@@ -343,16 +624,46 @@ describe("DNS Validated Certificate Handler", () => {
       });
   });
 
-  test("Delete operation deletes the certificate", () => {
+  test("Delete operation deletes the last legacy cert", () => {
     const describeCertificateFake = sinon.fake.resolves({
       Certificate: {
         CertificateArn: testCertificateArn,
+        DomainValidationOptions: legacyCertValidatorOptions,
       },
     });
     AWS.mock("ACM", "describeCertificate", describeCertificateFake);
 
     const deleteCertificateFake = sinon.fake.resolves({});
     AWS.mock("ACM", "deleteCertificate", deleteCertificateFake);
+
+    const listCertificatesFake = sinon.stub();
+    listCertificatesFake.onFirstCall().resolves({
+      CertificateSummaryList: [
+        {
+          DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+        },
+      ],
+      NextToken: "some token",
+    });
+    listCertificatesFake.onSecondCall().resolves({
+      CertificateSummaryList: [
+        {
+          DomainName: "otherdomain.com",
+        },
+      ],
+    });
+    AWS.mock("ACM", "listCertificates", listCertificatesFake);
+
+    const changeResourceRecordSetsFake = sinon.fake.resolves({
+      ChangeInfo: {
+        Id: "bogus",
+      },
+    });
+    AWS.mock(
+      "Route53",
+      "changeResourceRecordSets",
+      changeResourceRecordSetsFake
+    );
 
     const request = nock(ResponseURL)
       .put("/", (body) => {
@@ -366,8 +677,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
-          HostedZoneId: testHostedZoneId,
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
@@ -375,6 +690,295 @@ describe("DNS Validated Certificate Handler", () => {
           describeCertificateFake,
           sinon.match({
             CertificateArn: testCertificateArn,
+          })
+        );
+        sinon.assert.calledWith(listCertificatesFake, sinon.match({}));
+        sinon.assert.calledWith(
+          listCertificatesFake,
+          sinon.match({
+            NextToken: "some token",
+          })
+        );
+        sinon.assert.calledWith(
+          changeResourceRecordSetsFake,
+          sinon.match({
+            ChangeBatch: testDeleteRecordChangebatch,
+            HostedZoneId: testHostedZoneId,
+          })
+        );
+        sinon.assert.calledWith(
+          deleteCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        expect(request.isDone()).toBe(true);
+      });
+  });
+
+  test("Delete operation deletes the legacy cert but not the last", () => {
+    const describeCertificateFake = sinon.fake.resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+        DomainValidationOptions: legacyCertValidatorOptions,
+      },
+    });
+    AWS.mock("ACM", "describeCertificate", describeCertificateFake);
+
+    const deleteCertificateFake = sinon.fake.resolves({});
+    AWS.mock("ACM", "deleteCertificate", deleteCertificateFake);
+
+    const listCertificatesFake = sinon.stub();
+    listCertificatesFake.resolves({
+      CertificateSummaryList: [
+        {
+          DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+        },
+        {
+          DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+        },
+      ],
+    });
+    AWS.mock("ACM", "listCertificates", listCertificatesFake);
+
+    const request = nock(ResponseURL)
+      .put("/", (body) => {
+        return body.Status === "SUCCESS";
+      })
+      .reply(200);
+
+    return LambdaTester(handler.certificateRequestHandler)
+      .event({
+        RequestType: "Delete",
+        RequestId: testRequestId,
+        PhysicalResourceId: testCertificateArn,
+        ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
+          Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+        },
+      })
+      .expectResolve(() => {
+        sinon.assert.calledWith(
+          describeCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        sinon.assert.calledWith(listCertificatesFake, sinon.match({}));
+        sinon.assert.calledWith(
+          deleteCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        expect(request.isDone()).toBe(true);
+      });
+  });
+
+  test("Delete operation deletes the last new cert", () => {
+    const describeCertificateFake = sinon.fake.resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+        DomainValidationOptions: newCertValidateOptions,
+      },
+    });
+    AWS.mock("ACM", "describeCertificate", describeCertificateFake);
+
+    const deleteCertificateFake = sinon.fake.resolves({});
+    AWS.mock("ACM", "deleteCertificate", deleteCertificateFake);
+
+    const listCertificatesFake = sinon.stub();
+    listCertificatesFake.resolves({
+      CertificateSummaryList: [
+        {
+          DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+        },
+      ],
+    });
+    AWS.mock("ACM", "listCertificates", listCertificatesFake);
+
+    const changeResourceRecordSetsFake = sinon.fake.resolves({
+      ChangeInfo: {
+        Id: "bogus",
+      },
+    });
+    AWS.mock(
+      "Route53",
+      "changeResourceRecordSets",
+      changeResourceRecordSetsFake
+    );
+
+    const listHostedZonesByNameFake = sinon.fake.resolves({
+      HostedZones: [
+        {
+          Id: `/hostedzone/${testAppHostedZoneId}`,
+        },
+      ],
+    });
+    AWS.mock("Route53", "listHostedZonesByName", listHostedZonesByNameFake);
+
+    const request = nock(ResponseURL)
+      .put("/", (body) => {
+        return body.Status === "SUCCESS";
+      })
+      .reply(200);
+
+    return LambdaTester(handler.certificateRequestHandler)
+      .event({
+        RequestType: "Delete",
+        RequestId: testRequestId,
+        PhysicalResourceId: testCertificateArn,
+        ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
+          Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+        },
+      })
+      .expectResolve(() => {
+        sinon.assert.calledWith(
+          describeCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        sinon.assert.calledWith(listCertificatesFake, sinon.match({}));
+        sinon.assert.calledWith(
+          changeResourceRecordSetsFake,
+          sinon.match({
+            ChangeBatch: testDeleteRecordChangebatch,
+            HostedZoneId: testHostedZoneId,
+          })
+        );
+        sinon.assert.calledWith(
+          changeResourceRecordSetsFake,
+          sinon.match({
+            ChangeBatch: testDeleteRecordChangebatch,
+            HostedZoneId: testAppHostedZoneId,
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testAppName}.${testDomainName}`,
+            MaxItems: "1",
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testDomainName}`,
+            MaxItems: "1",
+          })
+        );
+        sinon.assert.calledWith(
+          deleteCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        expect(request.isDone()).toBe(true);
+      });
+  });
+
+  test("Delete operation deletes the last new cert but not the last", () => {
+    const describeCertificateFake = sinon.fake.resolves({
+      Certificate: {
+        CertificateArn: testCertificateArn,
+        DomainValidationOptions: newCertValidateOptions,
+      },
+    });
+    AWS.mock("ACM", "describeCertificate", describeCertificateFake);
+
+    const deleteCertificateFake = sinon.fake.resolves({});
+    AWS.mock("ACM", "deleteCertificate", deleteCertificateFake);
+
+    const listCertificatesFake = sinon.stub();
+    listCertificatesFake.resolves({
+      CertificateSummaryList: [
+        {
+          DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+        },
+        {
+          DomainName: `${testEnvName}.${testAppName}.${testDomainName}`,
+        },
+      ],
+    });
+    AWS.mock("ACM", "listCertificates", listCertificatesFake);
+
+    const changeResourceRecordSetsFake = sinon.fake.resolves({
+      ChangeInfo: {
+        Id: "bogus",
+      },
+    });
+    AWS.mock(
+      "Route53",
+      "changeResourceRecordSets",
+      changeResourceRecordSetsFake
+    );
+
+    const listHostedZonesByNameFake = sinon.fake.resolves({
+      HostedZones: [
+        {
+          Id: `/hostedzone/${testAppHostedZoneId}`,
+        },
+      ],
+    });
+    AWS.mock("Route53", "listHostedZonesByName", listHostedZonesByNameFake);
+
+    const request = nock(ResponseURL)
+      .put("/", (body) => {
+        return body.Status === "SUCCESS";
+      })
+      .reply(200);
+
+    return LambdaTester(handler.certificateRequestHandler)
+      .event({
+        RequestType: "Delete",
+        RequestId: testRequestId,
+        PhysicalResourceId: testCertificateArn,
+        ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
+          Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
+        },
+      })
+      .expectResolve(() => {
+        sinon.assert.calledWith(
+          describeCertificateFake,
+          sinon.match({
+            CertificateArn: testCertificateArn,
+          })
+        );
+        sinon.assert.calledWith(listCertificatesFake, sinon.match({}));
+        sinon.assert.calledWith(
+          changeResourceRecordSetsFake,
+          sinon.match({
+            ChangeBatch: testDeleteRecordChangebatch,
+            HostedZoneId: testAppHostedZoneId,
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testAppName}.${testDomainName}`,
+            MaxItems: "1",
+          })
+        );
+        sinon.assert.calledWith(
+          listHostedZonesByNameFake,
+          sinon.match({
+            DNSName: `${testDomainName}`,
+            MaxItems: "1",
           })
         );
         sinon.assert.calledWith(
@@ -409,7 +1013,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
@@ -438,16 +1047,6 @@ describe("DNS Validated Certificate Handler", () => {
       Certificate: {
         CertificateArn: testCertificateArn,
         InUseBy: [usedByArn],
-        DomainValidationOptions: [
-          {
-            ValidationStatus: "SUCCESS",
-            ResourceRecord: {
-              Name: testRRName,
-              Type: "CNAME",
-              Value: testRRValue,
-            },
-          },
-        ],
       },
     });
 
@@ -455,29 +1054,8 @@ describe("DNS Validated Certificate Handler", () => {
       Certificate: {
         CertificateArn: testCertificateArn,
         InUseBy: [],
-        DomainValidationOptions: [
-          {
-            ValidationStatus: "SUCCESS",
-            ResourceRecord: {
-              Name: testRRName,
-              Type: "CNAME",
-              Value: testRRValue,
-            },
-          },
-        ],
       },
     });
-
-    const changeResourceRecordSetsFake = sinon.fake.resolves({
-      ChangeInfo: {
-        Id: "bogus",
-      },
-    });
-    AWS.mock(
-      "Route53",
-      "changeResourceRecordSets",
-      changeResourceRecordSetsFake
-    );
     AWS.mock("ACM", "describeCertificate", describeCertificateFake);
 
     const deleteCertificateFake = sinon.fake.resolves({});
@@ -495,8 +1073,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
-          HostedZoneId: testHostedZoneId,
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
@@ -510,29 +1092,6 @@ describe("DNS Validated Certificate Handler", () => {
           deleteCertificateFake,
           sinon.match({
             CertificateArn: testCertificateArn,
-          })
-        );
-        sinon.assert.calledWith(
-          changeResourceRecordSetsFake,
-          sinon.match({
-            ChangeBatch: {
-              Changes: [
-                {
-                  Action: "DELETE",
-                  ResourceRecordSet: {
-                    Name: testRRName,
-                    Type: "CNAME",
-                    TTL: 60,
-                    ResourceRecords: [
-                      {
-                        Value: testRRValue,
-                      },
-                    ],
-                  },
-                },
-              ],
-            },
-            HostedZoneId: testHostedZoneId,
           })
         );
         expect(request.isDone()).toBe(true);
@@ -568,8 +1127,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
-          HostedZoneId: testHostedZoneId,
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
@@ -617,8 +1180,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
-          HostedZoneId: testHostedZoneId,
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
@@ -663,8 +1230,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
-          HostedZoneId: testHostedZoneId,
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
@@ -708,8 +1279,12 @@ describe("DNS Validated Certificate Handler", () => {
         RequestId: testRequestId,
         PhysicalResourceId: testCertificateArn,
         ResourceProperties: {
-          HostedZoneId: testHostedZoneId,
+          AppName: testAppName,
+          EnvName: testEnvName,
+          DomainName: testDomainName,
+          EnvHostedZoneId: testHostedZoneId,
           Region: "us-east-1",
+          RootDNSRole: testRootDNSRole,
         },
       })
       .expectResolve(() => {
