+++ templates/environment/partials/custom-resources.yml
@@ -30,17 +30,17 @@ HTTPSCert:
   - DelegateDNSAction
   Properties:
     ServiceToken: !GetAtt CertificateValidationFunction.Arn
-    DomainName: !Sub ${EnvironmentName}.${AppName}.${AppDNSName}
     HostedZoneId: !Ref EnvironmentHostedZone
-    # AppName: !Ref AppName
-    # EnvName: !Ref EnvironmentName
-    # DomainName: !Ref AppDNSName
-    # EnvHostedZoneId: !Ref EnvironmentHostedZone
+    AppName: !Ref AppName
+    EnvName: !Ref EnvironmentName
+    DomainName: !Ref AppDNSName
+    EnvHostedZoneId: !Ref EnvironmentHostedZone
     Region: !Ref AWS::Region
-    # RootDNSRole: !Ref AppDNSDelegationRole
+    RootDNSRole: !Ref AppDNSDelegationRole
+    IsAliasEnabled: !If [HasAliases, "true", "false"]
     SubjectAlternativeNames:
-    # - !Sub "${AppDNSName}"
-    # - !Sub "*.${AppDNSName}"
-    # - !Sub "${AppName}.${AppDNSName}"
-    # - !Sub "*.${AppName}.${AppDNSName}"
+    - !Sub "${AppDNSName}"
+    - !Sub "*.${AppDNSName}"
+    - !Sub "${AppName}.${AppDNSName}"
+    - !Sub "*.${AppName}.${AppDNSName}"
     - !Sub "*.${EnvironmentName}.${AppName}.${AppDNSName}"
